requirements.txt
Django==1.9.7
argparse==1.2.1
django-environ==0.4.0
djangorestframework==3.3.3
six==1.10.0
wsgiref==0.1.2
